+++ IRCConnectionTest/Events/CustomEventArgs/MessageEventArgs.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace IRCConnectionTest.Events.ComstumEventArgs
+namespace IRCConnectionTest.Events.CustomEventArgs
 {
     internal class MessageEventArgs : EventArgs
     {
