+++ IRCConnectionTest/Events/CustomEventArgs/RoomStateSlowModeEventArgs.cs
@@ -1,4 +1,4 @@
-﻿namespace IRCConnectionTest.Events.ComstumEventArgs
+﻿namespace IRCConnectionTest.Events.CustomEventArgs
 {
     internal class RoomStateSlowModeEventArgs
     {
