+++ IRCConnectionTest/Events/UserEventManager.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Text.RegularExpressions;
-using IRCConnectionTest.Events.ComstumEventArgs;
+using IRCConnectionTest.Events.CustomEventArgs;
 using IRCConnectionTest.Events.Misc;
 using IRCConnectionTest.Misc;
