+++ IRCConnectionTest/IrcConnection.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Net.Sockets;
 using System.Threading;
-using IRCConnectionTest.Events.ComstumEventArgs;
+using IRCConnectionTest.Events.CustomEventArgs;
 using IRCConnectionTest.Misc;
 
 namespace IRCConnectionTest
