+++ IRCConnectionTest/UserList.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Timers;
 using IRCConnectionTest.Events;
-using IRCConnectionTest.Events.ComstumEventArgs;
+using IRCConnectionTest.Events.CustomEventArgs;
 using IRCConnectionTest.Misc;
 using IRCConnectionTest.TmiApi.ChannelChattersEndpoint;
 using Timer = System.Timers.Timer;
